refactor: improve thread handling and damage calculation logic in autoAnchor.java; clean up unused code and imports
- improvements

(going away in 1.5 weeks)

+++ src/main/java/dev/seasnail1/modules/render/FOV.java
@@ -1,15 +1,24 @@
 package dev.seasnail1.modules.render;
 
 
+import com.jcraft.jorbis.Block;
 import dev.seasnail1.Addon;
+import dev.seasnail1.utilities.MathHelper;
 import meteordevelopment.meteorclient.events.render.GetFovEvent;
+import meteordevelopment.meteorclient.events.render.Render3DEvent;
+import meteordevelopment.meteorclient.renderer.ShapeMode;
 import meteordevelopment.meteorclient.settings.IntSetting;
 import meteordevelopment.meteorclient.settings.Setting;
 import meteordevelopment.meteorclient.settings.SettingGroup;
 import meteordevelopment.meteorclient.systems.modules.Module;
+import meteordevelopment.meteorclient.utils.render.color.Color;
 import meteordevelopment.orbit.EventHandler;
+import net.minecraft.util.math.BlockPos;
 
 import java.util.Arrays;
+import java.util.List;
+
+import static dev.seasnail1.utilities.MathHelper.getHoles;
 
 public class FOV extends Module {
     private final SettingGroup sgGeneral = settings.getDefaultGroup();
@@ -29,7 +38,7 @@ public class FOV extends Module {
     @EventHandler
     private void FOVModify(GetFovEvent event) {
         try {
-            event.fov = Fov.get();
+            event.fov = Fov.get().floatValue();
         } catch (Exception e) {
             info("An error occurred while changing the FOV");
             Addon.Logger.error("An error occurred while changing the FOV  {}", Arrays.toString(e.getStackTrace()));
